+++ src/battle_pike.c
@@ -811,7 +811,7 @@ static void HealMon(struct Pokemon *mon)
 
 static bool8 DoesAbilityPreventStatus(struct Pokemon *mon, u32 status)
 {
-    u16 ability = GetMonAbility(mon);
+    u8 ability = GetMonAbility(mon);
     bool8 ret = FALSE;
 
     switch (status)
